Move QT6 headers from app/qt -> include/qt
