+++ app/qt/CMakeLists.txt
@@ -1,3 +1,6 @@
+# Include QT header files
+set(QT_HEADERS ${PROJECT_SOURCE_DIR}/include/qt)
+
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 set(CMAKE_AUTOUIC ON)
@@ -19,18 +22,25 @@ set(CMAKE_AUTORCC ON)
 
 find_package(Qt6 REQUIRED COMPONENTS Widgets)
 
+include_directories(${QT_HEADERS})
+
 # Compile the needed individual components
 add_subdirectory("datamodel")
 add_subdirectory("embedwidget")
 add_subdirectory("mainwindow")
 
-set(QTUI_HEADERS_DIR "${PROJECT_SOURCE_DIR}/app/qt")
-set(HEADER_LIST "${QTUI_HEADERS_DIR}/embedwidget" "${QTUI_HEADERS_DIR}/mainwindow")
 
 # Build the main executable
 set(TARGET oggex-qt)
 qt_add_executable(${TARGET} main.cpp)
 
+#qt6_wrap_cpp(mainwindow/mainwindow.cpp ${QT_HEADERS}/mainwindow.h)
+#qt6_wrap_ui(mainwindow/mainwindow.ui mainwindow/mainwindow.ui)
+
+#qt6_wrap_cpp(embedwidget/embedwidget.cpp ${QT_HEADERS}/embedwidget.h)
+#qt6_wrap_ui(embedwidget/embedwidget.ui embedwidget/embedwidget.ui)
+
 # Include necessary header files & link components
-target_include_directories(${TARGET} PUBLIC ${HEADER_LIST})
+#target_include_directories(${TARGET} PUBLIC ${QT_HEADERS} ./mainwindow ./embedwidget ./datamodel)
+target_include_directories(${TARGET} PUBLIC ${QT_HEADERS})
 target_link_libraries(${TARGET} PRIVATE Qt6::Widgets embedwidget mainwindow)
